# openapi.yaml
openapi: 3.0.3
info:
  title: tileserve-go
  description: >
    A small Go HTTP server for managing maps and their versioned tile uploads,
    backed by Postgres and secured with JWT authentication. Maps have
    versioned zip uploads extracted to `<data-root>/<uuid>/<version>/` on
    disk, served back at `/maps/{id}/version/{version}/...`.
  version: "1.0.0"
servers:
  - url: http://localhost:8085
...
security:
  - bearerAuth: []
  - tokenQuery: []
paths:
...
  /maps:
    get:
      tags: [maps]
      summary: List maps
      description: >-
        Returns only the maps the acting user may see: maps marked
        `visibleToAll`, maps they created, maps they hold a per-map
        view/edit/delete grant on, admins, and users whose global `can_edit`
        or `can_delete` permission already lets them act on every map.
      responses:
        "200":
          description: Array of visible maps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Map"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [maps]
      summary: Create a map
      description: Requires the acting user's `can_create` permission.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MapRequest"
      responses:
        "201":
          description: Map created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Map"
        "400":
          description: Invalid request body or missing name
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /maps/{id}:
    parameters:
      - $ref: "#/components/parameters/MapId"
    get:
      tags: [maps]
      summary: Get a map
      description: Requires view access to this map (see `GET /maps`); returns 403 if the map is private and the acting user can't see it.
      responses:
        "200":
          description: The map
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Map"
        "400":
          description: Invalid map id
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [maps]
      summary: Replace a map's name/currentVersion/visibleToAll/anonymousAllowed
      description: >
        Requires the acting user's `can_edit` permission (global or per-map).
        `name`, `currentVersion`, `visibleToAll`, and `anonymousAllowed` are
        replaced; this does not touch the map_versions history or any
        extracted files on disk.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MapRequest"
      responses:
        "200":
          description: Updated map
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Map"
        "400":
          description: Invalid request body or missing name
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [maps]
      summary: Delete a map
      description: Requires the acting user's `can_delete` permission. Cascades and deletes the map's version history.
      responses:
        "204":
          description: Map deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    tokenQuery:
      type: apiKey
      in: query
      name: token
      description: JWT passed as a query parameter instead of the Authorization header.
  parameters:
    MapId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Unauthorized:
      description: Missing, invalid, or expired token
      content:
        text/plain:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
    Forbidden:
      description: Acting user lacks the required permission
      content:
        text/plain:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
    NotFound:
      description: Map not found
      content:
        text/plain:
          schema:
            $ref: "#/components/schemas/ErrorMessage"
  schemas:
    ErrorMessage:
      type: string
      description: Plain-text error message.
...
    MapRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        currentVersion:
          type: string
        visibleToAll:
          type: boolean
          description: Whether every authenticated user can see this map. Defaults to false (private) if omitted.
        anonymousAllowed:
          type: boolean
          description: >-
            Whether GET /maps/{id}/version/{version}/* (raw tile files) may
            be fetched with no bearer token at all. Independent of
            `visibleToAll` — a map can be private to authenticated users yet
            still serve tiles anonymously, or vice versa. Defaults to false
            if omitted.
    Map:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        name:
          type: string
        currentVersion:
          type: string
        visibleToAll:
          type: boolean
          description: >-
            Whether every authenticated user can see this map. When false,
            only admins, the creator, users whose global `can_edit`/`can_delete`
            already lets them act on any map, and users with a per-map grant
            can see it.
        anonymousAllowed:
          type: boolean
          description: >-
            Whether this map's tile files can be fetched with no bearer
            token at all (see `GET /maps/{id}/version/{version}/{filePath}`).
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdBy:
          type: string
        updatedBy:
          type: string
...